Rename v10 modules and noise to crypto
