+++ src/lib.rs
@@ -88,27 +88,23 @@
 mod builder;
 mod channels;
 mod constants;
+mod crypto;
 mod duplex;
 mod message;
-mod message_v10;
-mod noise;
 mod protocol;
 mod reader;
 mod util;
 mod writer;
 
 /// The wire messages used by the protocol.
-pub mod schema {
-    include!("schema_v10.rs");
-}
+pub mod schema;
 
 pub use builder::{Builder as ProtocolBuilder, Options};
-
 pub use channels::Channel;
 // Export the needed types for Channel::take_receiver, and Channel::local_sender()
 pub use async_channel::{Receiver as ChannelReceiver, Sender as ChannelSender};
 pub use duplex::Duplex;
 pub use hypercore; // Re-export hypercore
-pub use message_v10::Message;
+pub use message::Message;
 pub use protocol::{DiscoveryKey, Event, Key, Protocol};
 pub use util::discovery_key;
